+++ asdf/tags/wcs/tests/test_wcs.py
@@ -3,36 +3,23 @@
 
 from __future__ import absolute_import, division, unicode_literals, print_function
 
-
-try:
-    import astropy
-except ImportError:
-    HAS_ASTROPY = False
-else:
-    HAS_ASTROPY = True
-
-    from astropy.modeling import models
-    from astropy import coordinates as coord
-    from astropy import units as u
-    from astropy import time
-
-try:
-    import gwcs
-except ImportError:
-    HAS_GWCS = False
-else:
-    HAS_GWCS = True
-
-    from gwcs import coordinate_frames as cf
-    from gwcs import wcs
-
 import pytest
 import warnings
+
+gwcs = pytest.importorskip('gwcs')
+astropy = pytest.importorskip('astropy', minversion='1.3.3')
+
+from astropy.modeling import models
+from astropy import coordinates as coord
+from astropy import units as u
+from astropy import time
+
+from gwcs import coordinate_frames as cf
+from gwcs import wcs
 
 from ....tests import helpers
 
 
-@pytest.mark.skipif('not HAS_GWCS')
 def test_create_wcs(tmpdir):
     m1 = models.Shift(12.4) & models.Shift(-2)
     m2 = models.Scale(2) & models.Scale(-2)
@@ -51,7 +38,6 @@ def test_create_wcs(tmpdir):
     helpers.assert_roundtrip_tree(tree, tmpdir)
 
 
-@pytest.mark.skipif('not HAS_GWCS')
 def test_composite_frame(tmpdir):
     icrs = coord.ICRS()
     fk5 = coord.FK5()
@@ -73,9 +59,8 @@ def test_composite_frame(tmpdir):
 
     helpers.assert_roundtrip_tree(tree, tmpdir)
 
-
-@pytest.mark.skipif('not HAS_GWCS')
-def test_frames(tmpdir):
+def create_test_frames():
+    """Creates an array of frames to be used for testing."""
 
     # Suppress warnings from astropy that are caused by having 'dubious' dates
     # that are too far in the future. It's not a concern for the purposes of
@@ -106,18 +91,17 @@ def test_frames(tmpdir):
 
         cf.CelestialFrame(
             reference_frame=coord.Galactocentric(
+                # A default galcen_coord is used since none is provided here
                 galcen_distance=5.0*u.m,
-                galcen_ra=45*u.deg,
-                galcen_dec=1*u.rad,
                 z_sun=3*u.pc,
                 roll=3*u.deg)
             ),
 
-        #cf.CelestialFrame(
-        #    reference_frame=coord.GCRS(
-        #        obstime=time.Time('2010-01-01'),
-        #        obsgeoloc=[1, 3, 2000] * u.pc,
-        #        obsgeovel=[2, 1, 8] * (u.m/u.s))),
+        cf.CelestialFrame(
+            reference_frame=coord.GCRS(
+                obstime=time.Time('2010-01-01'),
+                obsgeoloc=[1, 3, 2000] * u.pc,
+                obsgeovel=[2, 1, 8] * (u.m/u.s))),
 
         cf.CelestialFrame(
             reference_frame=coord.CIRS(
@@ -127,15 +111,19 @@ def test_frames(tmpdir):
             reference_frame=coord.ITRS(
                 obstime=time.Time('2022-01-03'))),
 
-        #cf.CelestialFrame(
-        #    reference_frame=coord.PrecessedGeocentric(
-        #        obstime=time.Time('2010-01-01'),
-        #        obsgeoloc=[1, 3, 2000] * u.pc,
-        #        obsgeovel=[2, 1, 8] * (u.m/u.s)))
+        cf.CelestialFrame(
+            reference_frame=coord.PrecessedGeocentric(
+                obstime=time.Time('2010-01-01'),
+                obsgeoloc=[1, 3, 2000] * u.pc,
+                obsgeovel=[2, 1, 8] * (u.m/u.s)))
     ]
 
+    return frames
+
+def test_frames(tmpdir):
+
     tree = {
-        'frames': frames
+        'frames': create_test_frames()
     }
 
     helpers.assert_roundtrip_tree(tree, tmpdir)