+++ CHANGES.rst
@@ -7,11 +7,18 @@
 
 - Added a function ``is_asdf_file`` which inspects the input and
   returns ``True`` or ``False``. [#239]
-
+  
 - The ``open`` method of ``AsdfInFits`` now accepts URIs and open file handles
   in addition to HDULists. The ``open`` method of ``AsdfFile`` will now try to
   parse the given URI or file handle as ``AsdfInFits`` if it is not obviously a
   regular ASDF file. [#241]
+  
+- Updated WCS frame fields ``obsgeoloc`` and ``obsgeovel`` to reflect recent
+  updates in ``astropy`` that changed representation from ``Quantity`` to
+  ``CartesianRepresentation``. Updated to reflect ``astropy`` change that
+  combines ``galcen_ra`` and ``galcen_dec`` into ``galcen_coord``. Added
+  support for new field ``galcen_v_sun``. Added support for required module
+  versions for tag classes. [#244]
 
 1.2.1(2016-11-07)
 -----------------